drivers/net/xen: module refacroring
