+++ templates/xen/debug/mods.config
@@ -36,5 +36,5 @@ configuration conf {
 	include embox.cmd.sys.date
 	include embox.cmd.msleep
 
-	include embox.driver.net.xen_pv_net
+	include embox.driver.net.xen
 }
